+++ src/main/java/hello/cluebackend/domain/notice/controller/dto/response/NoticeDto.java
@@ -1,6 +1,6 @@
 package hello.cluebackend.domain.notice.controller.dto.response;
 
-import hello.cluebackend.domain.notice.domain.NoticeType;
+import hello.cluebackend.domain.notice.model.NoticeType;
 import lombok.*;
 
 import java.time.LocalDateTime;
