+++ src/main/java/hello/cluebackend/domain/notice/model/NoticeType.java
@@ -1,4 +1,4 @@
-package hello.cluebackend.domain.notice.domain;
+package hello.cluebackend.domain.notice.model;
 
 public enum NoticeType {
     SCHOOL, SCHEDULE, SERVICE
