+++ src/main/java/hello/cluebackend/presentation/api/notice/NoticeCommandController.java
@@ -1,31 +1,24 @@
-package hello.cluebackend.domain.notice.controller;
+package hello.cluebackend.presentation.api.notice;
 
 import hello.cluebackend.application.user.dto.CustomOAuth2User;
 import hello.cluebackend.domain.notice.service.NoticeService;
 import hello.cluebackend.domain.notice.controller.dto.request.AddNoticeDto;
 import hello.cluebackend.domain.notice.controller.dto.request.CreateNoticeDto;
 import hello.cluebackend.domain.notice.controller.dto.request.ModifyNoticeDto;
-import hello.cluebackend.domain.notice.controller.dto.response.NoticeDto;
-import hello.cluebackend.domain.notice.controller.dto.response.NoticeInfoDto;
-import hello.cluebackend.domain.noticedocument.controller.dto.response.NoticeDownloadDto;
-import hello.cluebackend.domain.noticedocument.controller.dto.response.NoticeUrlDto;
 import lombok.RequiredArgsConstructor;
-import org.springframework.core.io.Resource;
 import org.springframework.http.*;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.UUID;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/notice")
-public class NoticeController {
+public class NoticeCommandController {
 
     private final NoticeService noticeService;
 
@@ -73,45 +66,6 @@ public class NoticeController {
         UUID userId = customOAuth2User.getUserDTO().getUserId();
         noticeService.remove(userId, noticeId);
         return ResponseEntity.status(HttpStatus.OK).build();
-    }
-
-    @GetMapping
-    public ResponseEntity<List<NoticeDto>> getAllNotices(@AuthenticationPrincipal CustomOAuth2User customOAuth2User) {
-        UUID userId = customOAuth2User.getUserDTO().getUserId();
-        return ResponseEntity.status(HttpStatus.OK).body(noticeService.findAllById());
-    }
-
-    @GetMapping("/{noticeId}")
-    public ResponseEntity<NoticeInfoDto> getNotice(
-            @PathVariable UUID noticeId,
-            @AuthenticationPrincipal CustomOAuth2User customOAuth2User) {
-        UUID userId = customOAuth2User.getUserDTO().getUserId();
-        return ResponseEntity.status(HttpStatus.OK).body(noticeService.findById(userId, noticeId));
-    }
-
-    @GetMapping("/download/{noticeDocumentId}")
-    public ResponseEntity<Resource> downloadNoticeDocument(
-            @PathVariable("noticeDocumentId")  UUID noticeDocumentId) throws IOException {
-        NoticeDownloadDto dto = noticeService.downloadNoticeDocument(noticeDocumentId);
-
-        String original = dto.getOriginal();
-        String contentType = dto.getContentType();
-        MediaType mediaType = (contentType != null) ? MediaType.parseMediaType(contentType) : MediaType.ALL;
-
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .contentType(mediaType)
-                .header(HttpHeaders.CONTENT_DISPOSITION, ContentDisposition.attachment()
-                        .filename(original, StandardCharsets.UTF_8)
-                        .build()
-                        .toString())
-                .body(dto.getResource());
-    }
-
-    @GetMapping("/link/{noticeDocumentId}")
-    public ResponseEntity<NoticeUrlDto> getNoticeLink(
-            @PathVariable("noticeDocumentId") UUID noticeDocumentId) {
-        return ResponseEntity.status(HttpStatus.OK).body(noticeService.getLink(noticeDocumentId));
     }
 
     @PreAuthorize("hasRole('TEACHER')")
